GET and POST working
